pyproject.toml
[tool.poetry]
name = "fanta-formazioni-bot"
version = "0.1.0"
description = "Telegram Bot to remind you of Fantacalcio's line-up"
authors = ["Matteo Pelliccione <mat.pelliccione@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mpellicc/fanta-formazioni-bot"
repository = "https://github.com/mpellicc/fanta-formazioni-bot"
keywords = ["telegram", "telegram-bot", "fantasy-football", "reminder", "notification"]

[tool.poetry.dependencies]
python = "3.12.4"
python-telegram-bot = {extras = ["job_queue"], version = "^21.4"}
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
ruff = "^0.5.7"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
